requirements.txt
nose==1.3.7
Pygments==2.2.0
python-fire==0.1.0
